Clean up, refactor, add sitemap

+++ application/config/scrabble.php
@@ -3,7 +3,7 @@
 return array(
 	'clubname'     => 'Winnipeg Scrabble Club',
 	'first_year'   => 2012,
-	'current_year' => date('Y'),
+	'current_year' => (int) date('Y'),
 	'current_wordlist' => 'wow24',
 
 	'initial_rating' => 1000
